Speed up the new broad band correction code.
We only measure the large scale power from a consistent
subsample for the particles on a consistent low resolution
mesh.

the examples are updated to use the smallest box that contains k ~ 0.02
modes. This is to minimize the power at mesh Nyquist which would have
become aliasing noise that affects the large scale power.