Fix a multitude of bugs in FOF.
two major bugs,

one is in assign halo attr, where the remote connection
is not properly detected. If a particle has ghost it still may connect
to a remote rank, thus the number of particle count is incomplete and
we cannot apply a filter yet.

the other is in the global merge. we cannot use send there in the iterations,
as it will miss changes due to updated ghosts. A proper reduction is needed.

Current API requires us to have this commbuff, but it can be removed (
together with task member) if we refactor the reduction function in
pmghosts to take any function as input. We can probably also also alo
a vector reduction on vector fields, then pack_member and reduce_member
are no longer needed.