Do not redistribute particles before IO.
For balanced particles, there is no need to redistribute.
For unbalanced particles, if we redistribute, then
almost all ranks participate in the sorting, and the aggregation
tends to not being able to be effective because the input
and output distribution are very different (mpsort could have
done better here). This has been observed with the KNL 4096 node runs.
Sorting of lightcone, even FOF halos took 100s of seconds.

If we do not redistribute, at worse the IO is no slower than
a full snapshot. On burst buffer this amounts to no time. On lustre
maybe bad. But access to a fast IO system during run
will be the new norm; e.g. Perlmutter has something similar as well.