Sync fof and rfof tests.
There is a bug. When I rerun the fof with particles in the top 300
active halos on the same ll, I only get 252 halos back.

The bug also shows up as changing the kdtree threshold changes
the fof result if not all particles are active. (hence here
I am settig threshold to 1.

This can probably be fixed from kdcount -- expose the python API
and fix it there!