Use Consistent laplacian in 2lpt and PM.
This eases reverting to the legacy fastpm 3_4 used in hidden-valley.
The swith to 1_4 kernel (for consistency with fastpm-python) has killed
a lot of less massive halos (>>10% at 100 particles), such that the hidden-valley2
is drastically different form hidden-valley1 if we do not fully revert to
the legacy 3_4 where lpt and pm both uses the sinc based laplacian.