LPT linear density field respects pm_nc_factor.
biwei has found using B=2 in LPT significantly improves quality
of halos at high redshift. This is consistent to other's observations.

Therefore we change LPT to use the pm_nc_factor value specified
in the parameter file.

To restore the old behavior of using B=1 for LPT,
 add a pm_nc_factor control point at a=0.0
for B=1, and a very small nonzero control point for B=2.

pm_nc_factor = {{0.0, 1}, {0.0001, 2}}

The tests are updated to use the old behavior.