Fixing 1.5 * Omega_cdm
Ages ago, the plan was to move 1.5 * Omga_cdm from these places
to gravity.c. We did this, but then rebased with master, and the
gravity.c factors were lost. As an easy fix I have just put back all the
1.5 * Omega_cdm's where they used to be before we tried moving anything
around. Then if we do actually want to mocve these factors ot gravity.c
we cna do so later.