Speed up the interpolation of the FD distribution
Previously we would re allocate, init and free the gsl_interp
object each time we needed to interpolate something.
This slowed down the code A LOT.

Now we alloc and init the gsl_interp objects at the start of code and
point to them in the cosmology object. Then we free at the end of the
simualtion.

Code could probably be neatened, but it works.