Introduced GrowthInfo Object
Previously code would recompute the ODE each time it wanted a growth
factor, or one of its derivatives. This was slow.

We now have a GrowthInfo object which stores all useful growth information at a
specific value of a. This speeds up computation when timestepping,
and setting ICs in pm2lpt. Note there are still some places that still
use the old function to compute the growth: io.c, horizon.c, and solver_growth.
Probably want to phase this out later, but it shouldn't slow things down
much because they're not used repeatedly as in timestepping.

Note this commit is messy as there are other things that have changed:

I have added lots of commented code for testing different ICs for the
growth ODE. This will be removed later.

I have changed omega_cdm to omega_m in a couple of places as necessary
when ncdm is around.