Incorporated old method (LCDM approx) for growth factors
Initially I had completely replaced the old FastPM method of assuming an
LCDM only approximation. I've now put this code back in and enabled the
user to choose whether to use this old code, o the new full ODE solver.

This means that if one were to run with:
growth_mode = LCDM
T_cmb = 0
m_ncdm = {}
The output should, in principle, be the same as the old FastPM code.
To test this I added tests to travis for growth_mode = LCDM.

Recall the lightcone and nbodykit tests differ by one halo in one of the
lines from each test when comparing the current master branch (which
assumes LCDM approx), and the cosmology branch using growth_mode=ODE. So
the tests are slightly different.

Note I have made kick->q2 consistent with old code for LCDM.

Also I've had to hack in old dD/da for LCDM runs:
In the new code with the growth ODE I use dD/da = f1 * D1 / a.
But for old LCDM approx code, f1 is approximated, and so this equation
is not used. So I have put in the old dD/da formula for use in LCDM
mode.