Corrected lightcone-ODE test numbers.
A while ago we merged with master, implementing the new lightcone code.
The LCDM test was changed in the master branch, but not the ODE mode, so
corrected this to make tests pass (number of particles in lightcone
should be same in CDM and ODE test, but the minor difference in number
of objects written is the same as before the merge).