The differentiation kernel for initial condition generation has been modified and is now the default
An option has been added to use the old kernel. With
the new one the power spectrum of the initial conditions
agrees up to the Nyquist frequency with the linear theory.
Hence, the issue of the lack of low-mass halos, which was
caused by the absence of power near the Nyquist frequency
in the initial conditions genereated by the old kernel
has been resolved.

Additionally, the default value of the kmax_over_nyquist
parameter has been increased to 0.666, to address
the issues with the clustering of the low mass haloes
in presence of fNL. Moreover this value also prevents
the effects the aliasing associated with the phi^2 term.